# Optimize `AssetId.parse` and cache `_normalizePath`