This simple optimization reduced the full build time of a project from 30 seconds to 24 seconds.

Original time:

```
Built with build_runner in 30s; wrote 31 outputs.
```

With the optimization:

```
Built with build_runner in 24s; wrote 31 outputs.
```

---

Statistics for the `_normalizePath` cache:

```
- Total calls: 16,828
- Computed: 4,723
- Uncached responses: 28%
- Cached responses: 72%
- Total strings in the cache: 4,723
- Total length of strings in the cache: 153,911
```

---

```text
Dart SDK version: 3.8.2 (stable) (Wed Jul 16 08:06:06 2025 -0700) on "macos_x64"
```

---

* [x] I’ve reviewed the contributor guide and applied the relevant portions to this PR.